Create namespace if it does not exist
I thought about this when I was adding a new stack for an existing
service and found myself debating between using chat to initialize
a repository for a 5th time (it's deployed to as many clusters)
just for namespace creation, or to do it manually and out-of-band.
Neither seemed like great options to me, so I figured, why couldn't
we just create the namespace on the fly if it doesn't exist?

I don't think the tradeoff of requiring a namespace to already
exist necessarily makes sense any longer, even though in the
majority of cases at Shopify it will.

The initializer already enforces that a namespace is present,
so we can guarantee this receives a non-empty value.

This makes kubernetes-deploy a little more plug-and-play with,
IMO, minimal drawbacks.