Fix unstable deploy result
It is unavoidably possible for the resources to go from fully up to not fully up between last poll and status output. This is a consequence of how Kubernetes works, and not necessarily indicative of a problem. Instead of restarting the polling in this case, we should stick to our original conclusion re:success and pass the deploy.