Add resource types for automatically provisioned persistent volumes to the priority sequence
If you want Kubernetes to automatically provision you a persistent
volume to use with a persistent volume claim, you need the persistent
volume resource to be created first, and if you are using a kubernetes
plugin to create the persistent volume (like the GCS provisioner), the
StorageClass that uses that plugin needs to be persisted first as well.

So, lets deploy those before we try to deploy any PersistentVolumes!

See
https://kubernetes.io/docs/concepts/storage/persistent-volumes/#storageclasses
for more information