Remove hacky test
activeDeadlineSeconds doesn't make sense on a deployment, and k8s now validates this, preventing this test's setup from working. This was a regression test trying to simulate dead pods hanging around as a result of evictions. Performing an actual eviction against minikube does not seem to leave the pods around. The value of this test is very low (we now ignore the old replicaset completely, and there is a test for the ability to run multiple revisions at once) and it is not obvious how to make it work without the activeDeadlineSeconds hack, so I'm removing it.