Fix "Running tasks at the beginning of a deploy" link in README.md
Summary
-------

The hyperlink "Running tasks at the beginning of a deploy" in the
high level feature listing of the README.md references a non-existent
anchor (#deploy-time-tasks). Looking through the history it seems like
this anchor never existed in any commits in the master history
chain. It would seem this link was always intended to point to the
content listed in the section titled "Running tasks at the beginning
of a deploy", but was perhaps missed in renaming edit.

This change makes the hyperlink in the feature listing mirror that in
the of the kubernetes-deploy 'usage' table of contents.

References
----------

- Originating Commit https://github.com/Shopify/kubernetes-deploy/pull/120/commits/36d916c64ee20a24e694fca9b1961482c71217d6