Add a kubernetes-render command, which renders templates to STDOUT
This introduces kubernetes-render, which is a UNIX-philosphy abiding simple tool that can pipe ERB-rendered raw YAML along to the next place anyone wants to send it. It renders one template in a template_dir with bindings.

It does this by:

Adding a new RenderTask that invokes a Renderer on a passed template
Adding a new exe file that invokes RenderTask