Always set a deployment_id, even if the current_sha isn't set
The deployment_id seems like it should always be present, even if current_sha isn't set. Before this change, deployment_id would be nil if current_sha was not set, which means predeployed pods like migrations and whatnot collide. It doesn't seem like a stretch to just use a random string as a deployment ID if the current deployment ID scheme is just using randomness too, so this adds that as a fallback, and adds a test!