Do not fail fast for sync cache issues
It seems kubernetes might have issues mounting the same secret/configmap in many pods that causes
the pod to be in a `CreateContainerConfigError` state and an error message `failed to sync %s
cache`.

This PR stop failing fast for that case and will let kubernetes to try to eventually run the
pod (after some time and a new pod is created we've seen this issue to be resolved by it self).