Optimize chance of users reading error messages
In our experience, it is common for the other elements of debug output (primarily pod logs) to be long enough that they fill the screen, pushing the part of the output that most needs to call attention passing into scrollback. It is also very common for users to not scroll back appropriately, assuming the important output about a failure will be at the bottom of the output. This moves the important information to the bottom to increase the chance that a user will read it.